+++ simplebot_howdoi.py
@@ -3,6 +3,7 @@
 import json
 
 import simplebot
+from deltachat import Message
 from howdoi.howdoi import howdoi
 from pkg_resources import DistributionNotFound, get_distribution
 from simplebot.bot import Replies
@@ -14,15 +15,28 @@ except DistributionNotFound:
     __version__ = "0.0.0.dev0-unknown"
 
 
+@simplebot.filter
+def query_filter(message: Message, replies: Replies) -> None:
+    """Send me a question in private to get an answer."""
+    if not message.chat.is_group():
+        _search(message.text, replies)
+
+
 @simplebot.command(name="/howdoi")
 def cmd_howdoi(payload: str, replies: Replies) -> None:
     """Instant coding answers.
 
     Example:
     /howdoi format date bash
+
+    You can also send me the question direclty in private.
     """
+    _search(payload, replies)
+
+
+def _search(query: str, replies: Replies) -> None:
     try:
-        res = json.loads(howdoi("{} -j".format(payload)))[0]
+        res = json.loads(howdoi("{} -j".format(query)))[0]
         replies.add(text="{}\n\n↗️ {}".format(res["answer"], res["link"]))
     except (Exception, SystemExit):  # noqa
         replies.add(text="❌ Something went wrong.")
@@ -37,3 +51,11 @@ class TestPlugin:
 
         msg = mocker.get_one_reply("/howdoi -h")
         assert "❌" in msg.text
+
+    def test_filter(self, mocker):
+        msg = mocker.get_one_reply("format date bash")
+        assert "↗" in msg.text
+
+        # filter should work only in private/direct chat
+        msgs = mocker.get_replies("format date bash", group="group1")
+        assert not msgs