+++ CHANGELOG.rst
@@ -1,8 +1,15 @@
 Changelog
 =========
 
+`Unreleased`_
+-------------
+
+- added filter to allow search directly in private 1:1 chats.
+
 1.0.0
 -----
 
 - initial release
 
+
+.. _Unreleased: https://github.com/adbenitez/simplebot_howdoi/compare/v1.0.0...HEAD